// lib/types.d.ts
export interface paths {
// ...
  "/systems/{systemSymbol}/waypoints/{waypointSymbol}/market": {
    /**
     * Get Market
     * @description Retrieve imports, exports and exchange data from a marketplace. Requires a waypoint that has the `Marketplace` trait to use.
     *
     * Send a ship to the waypoint to access trade good prices and recent transactions. Refer to the [Market Overview page](https://docs.spacetraders.io/game-concepts/markets) to gain better a understanding of the market in the game.
     */
    get: operations["get-market"];
    parameters: {
      path: {
        /** @description The system symbol */
        systemSymbol: string;
        /** @description The waypoint symbol */
        waypointSymbol: string;
      };
    };
  };
// ...
  "/my/ships": {
    /**
     * List Ships
     * @description Return a paginated list of all of ships under your agent's ownership.
     */
    get: operations["get-my-ships"];
    /**
     * Purchase Ship
     * @description Purchase a ship from a Shipyard. In order to use this function, a ship under your agent's ownership must be in a waypoint that has the `Shipyard` trait, and the Shipyard must sell the type of the desired ship.
     *
     * Shipyards typically offer ship types, which are predefined templates of ships that have dedicated roles. A template comes with a preset of an engine, a reactor, and a frame. It may also include a few modules and mounts.
     */
    post: operations["purchase-ship"];
  };
  "/my/ships/{shipSymbol}": {
    /**
     * Get Ship
     * @description Retrieve the details of a ship under your agent's ownership.
     */
    get: operations["get-my-ship"];
    parameters: {
      path: {
        /** @description The symbol of the ship. */
        shipSymbol: string;
      };
    };
  };
  "/my/ships/{shipSymbol}/cargo": {
    /**
     * Get Ship Cargo
     * @description Retrieve the cargo of a ship under your agent's ownership.
     */
    get: operations["get-my-ship-cargo"];
    parameters: {
      path: {
        /** @description The symbol of the ship. */
        shipSymbol: string;
      };
    };
  };
  "/my/ships/{shipSymbol}/orbit": {
    /**
     * Orbit Ship
     * @description Attempt to move your ship into orbit at its current location. The request will only succeed if your ship is capable of moving into orbit at the time of the request.
     *
     * Orbiting ships are able to do actions that require the ship to be above surface such as navigating or extracting, but cannot access elements in their current waypoint, such as the market or a shipyard.
     *
     * The endpoint is idempotent - successive calls will succeed even if the ship is already in orbit.
     */
    post: operations["orbit-ship"];
    parameters: {
      path: {
        /** @description The symbol of the ship. */
        shipSymbol: string;
      };
    };
  };
// ...
  "/my/ships/{shipSymbol}/dock": {
    /**
     * Dock Ship
     * @description Attempt to dock your ship at its current location. Docking will only succeed if your ship is capable of docking at the time of the request.
     *
     * Docked ships can access elements in their current location, such as the market or a shipyard, but cannot do actions that require the ship to be above surface such as navigating or extracting.
     *
     * The endpoint is idempotent - successive calls will succeed even if the ship is already docked.
     */
    post: operations["dock-ship"];
    parameters: {
      path: {
        /** @description The symbol of the ship. */
        shipSymbol: string;
      };
    };
  };
// ...
  "/my/ships/{shipSymbol}/extract": {
    /**
     * Extract Resources
     * @description Extract resources from a waypoint that can be extracted, such as asteroid fields, into your ship. Send an optional survey as the payload to target specific yields.
     *
     * The ship must be in orbit to be able to extract and must have mining equipments installed that can extract goods, such as the `Gas Siphon` mount for gas-based goods or `Mining Laser` mount for ore-based goods.
     *
     * The survey property is now deprecated. See the `extract/survey` endpoint for more details.
     */
    post: operations["extract-resources"];
    parameters: {
      path: {
        /** @description The ship symbol. */
        shipSymbol: string;
      };
    };
  };
// ...
  "/my/ships/{shipSymbol}/navigate": {
    /**
     * Navigate Ship
     * @description Navigate to a target destination. The ship must be in orbit to use this function. The destination waypoint must be within the same system as the ship's current location. Navigating will consume the necessary fuel from the ship's manifest based on the distance to the target waypoint.
     *
     * The returned response will detail the route information including the expected time of arrival. Most ship actions are unavailable until the ship has arrived at it's destination.
     *
     * To travel between systems, see the ship's Warp or Jump actions.
     */
    post: operations["navigate-ship"];
    parameters: {
      path: {
        /** @description The ship symbol. */
        shipSymbol: string;
      };
    };
  };
// ...
  "/my/ships/{shipSymbol}/sell": {
    /**
     * Sell Cargo
     * @description Sell cargo in your ship to a market that trades this cargo. The ship must be docked in a waypoint that has the `Marketplace` trait in order to use this function.
     */
    post: operations["sell-cargo"];
    parameters: {
      path: {
        /** @description Symbol of a ship. */
        shipSymbol: string;
      };
    };
  };
// ...
  "/my/ships/{shipSymbol}/refuel": {
    /**
     * Refuel Ship
     * @description Refuel your ship by buying fuel from the local market.
     *
     * Requires the ship to be docked in a waypoint that has the `Marketplace` trait, and the market must be selling fuel in order to refuel.
     *
     * Each fuel bought from the market replenishes 100 units in your ship's fuel.
     *
     * Ships will always be refuel to their frame's maximum fuel capacity when using this action.
     */
    post: operations["refuel-ship"];
    parameters: {
      path: {
        /** @description The ship symbol. */
        shipSymbol: string;
      };
    };
  };
// ...
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
// ...
    Market: {
      /** @description The symbol of the market. The symbol is the same as the waypoint where the market is located. */
      symbol: string;
      /** @description The list of goods that are exported from this market. */
      exports: components["schemas"]["TradeGood"][];
      /** @description The list of goods that are sought as imports in this market. */
      imports: components["schemas"]["TradeGood"][];
      /** @description The list of goods that are bought and sold between agents at this market. */
      exchange: components["schemas"]["TradeGood"][];
      /** @description The list of recent transactions at this market. Visible only when a ship is present at the market. */
      transactions?: components["schemas"]["MarketTransaction"][];
      /** @description The list of goods that are traded at this market. Visible only when a ship is present at the market. */
      tradeGoods?: components["schemas"]["MarketTradeGood"][];
    };
// ...
    Ship: {
      /** @description The globally unique identifier of the ship in the following format: `[AGENT_SYMBOL]-[HEX_ID]` */
      symbol: string;
      registration: components["schemas"]["ShipRegistration"];
      nav: components["schemas"]["ShipNav"];
      crew: components["schemas"]["ShipCrew"];
      frame: components["schemas"]["ShipFrame"];
      reactor: components["schemas"]["ShipReactor"];
      engine: components["schemas"]["ShipEngine"];
      cooldown: components["schemas"]["Cooldown"];
      /** @description Modules installed in this ship. */
      modules: components["schemas"]["ShipModule"][];
      /** @description Mounts installed in this ship. */
      mounts: components["schemas"]["ShipMount"][];
      cargo: components["schemas"]["ShipCargo"];
      fuel: components["schemas"]["ShipFuel"];
    };
// ...
    Waypoint: {
      symbol: components["schemas"]["WaypointSymbol"];
      type: components["schemas"]["WaypointType"];
      systemSymbol: components["schemas"]["SystemSymbol"];
      /** @description Relative position of the waypoint on the system's x axis. This is not an absolute position in the universe. */
      x: number;
      /** @description Relative position of the waypoint on the system's y axis. This is not an absolute position in the universe. */
      y: number;
      /** @description Waypoints that orbit this waypoint. */
      orbitals: components["schemas"]["WaypointOrbital"][];
      /** @description The symbol of the parent waypoint, if this waypoint is in orbit around another waypoint. Otherwise this value is undefined. */
      orbits?: string;
      faction?: components["schemas"]["WaypointFaction"];
      /** @description The traits of the waypoint. */
      traits: components["schemas"]["WaypointTrait"][];
      /** @description The modifiers of the waypoint. */
      modifiers?: components["schemas"]["WaypointModifier"][];
      chart?: components["schemas"]["Chart"];
      /** @description True if the waypoint is under construction. */
      isUnderConstruction: boolean;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: {
  };
  headers: never;
  pathItems: never;
}